+++ RazorSvelte/App/authorized-page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import Layout from "./shared/layout/main.svelte";
+    import Layout from "./shared/layout/default";
     import { getFromJson } from "./shared/config";
 
     let user = getFromJson<{name: string, email: string, timezone: string, timestamp: string}>("user");