+++ RazorSvelte/Pages/OffcanvasNav.cshtml.cs
@@ -1,3 +1,0 @@
-namespace RazorSvelte.Pages;
-
-public class OffcanvasNavModel : PageModel {}